Fix unit tests with py27
Fix building unit tests with py27.
