+++ .testr.conf
@@ -2,6 +2,8 @@
 test_command=OS_STDOUT_CAPTURE=${OS_STDOUT_CAPTURE:-1} \
              OS_STDERR_CAPTURE=${OS_STDERR_CAPTURE:-1} \
              OS_TEST_TIMEOUT=${OS_TEST_TIMEOUT:-60} \
-             ${PYTHON:-python} -m subunit.run discover -t ./ . $LISTOPT $IDOPTION
+             ${PYTHON:-python} -m subunit.run discover \
+             -t ./nova/tests/unit/virt/lxd/ ./nova/tests/unit/virt/lxd \
+             $LISTOPT $IDOPTION
 test_id_option=--load-list $IDFILE
 test_list_option=--list
