--- nova/tests/unit/virt/lxd/session/test_migrate.py
@@ -1,38 +1,0 @@
-# Copyright 2015 Canonical Ltd
-# All Rights Reserved.
-#
-#    Licensed under the Apache License, Version 2.0 (the "License"); you may
-#    not use this file except in compliance with the License. You may obtain
-#    a copy of the License at
-#
-#         http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS,
-#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-#    implied. See the License for the specific language governing
-#    permissions and limitations under the License.
-
-import ddt
-import mock
-
-from nova import test
-
-from nova_lxd.nova.virt.lxd import session
-from nova_lxd.tests import stubs
-
-
-@ddt.ddt
-class SessionMigrateTest(test.NoDBTestCase):
-
-    def setUp(self):
-        super(SessionMigrateTest, self).setUp()
-
-        """This is so we can mock out pylxd API calls."""
-        self.ml = stubs.lxd_mock()
-        lxd_patcher = mock.patch('pylxd.api.API',
-                                 mock.Mock(return_value=self.ml))
-        lxd_patcher.start()
-        self.addCleanup(lxd_patcher.stop)
-
-        self.session = session.LXDAPISession()
