+++ nova/tests/unit/virt/lxd/test_migrate.py
@@ -15,10 +15,9 @@
 
 import mock
 
+import nova.conf
 from nova import test
 from nova.virt import fake
-
-from oslo_config import cfg
 
 from nova.virt.lxd import config
 from nova.virt.lxd import migrate
@@ -26,9 +25,7 @@ from nova.virt.lxd import operations
 from nova.virt.lxd import session
 import stubs
 
-CONF = cfg.CONF
-CONF.import_opt('my_ip', 'nova.netconf')
-
+CONF = nova.conf.CONF
 
 class LXDTestContainerMigrate(test.NoDBTestCase):
