+++ tox.ini
@@ -18,6 +18,10 @@ deps = -r{toxinidir}/requirements.txt
        -egit+https://github.com/openstack/nova#egg=nova
 commands = ostestr {posargs}
 
+[testenv:py27]
+commands = /bin/cp -r {toxinidir}/nova/virt/lxd/ {toxinidir}/.tox/py27/src/nova/nova/virt/
+           python setup.py testr --slowest --testr-args='{posargs}'
+
 [testenv:pep8]
 commands = flake8
